fix: 嘗試修復 delete activity log 發生 npe

+++ api/Dataservices/DeviceActivityLogDataservice.cs
@@ -39,6 +39,11 @@ namespace Homo.IotApi
                 .Take(limit)
                 .ToList();
 
+            if (items.Count() == 0)
+            {
+                return null;
+            }
+
             items.ForEach(item =>
             {
                 dbContext.DeviceActivityLog.Remove(item);
